+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - `.decimals()` no longer prints an error message when used on an integer
 - `two_digit_to_decimal_form` outputs in math mode correctly
 - Axes are ticked more appropriately
+- `find_m` axes are scaled better when `m = 0`
 
 ## [0.12.3] - 2026-08-03